refactor: adding issues strategys

+++ src/main/java/net/dds/domain/CustomerRepository.java
@@ -3,6 +3,6 @@ package net.dds.domain;
 import net.dds.domain.customer.Customer;
 
 public interface CustomerRepository {
-    Customer findByDocumentNumber(Long documentNumber);
+    Customer findByDocumentNumber(Integer documentNumber);
     void save(Customer customer);
 }
